abort feature

+++ src/pytestemb/valid.py
@@ -42,6 +42,12 @@ class Valid:
         self.destroy    = self._nothing_
         self.case = []
         self.name = utils.get_script_name()
+        
+        self.aborted = False
+        
+        self.abort_fatal_mode = False
+        self.abort_fatal = False
+        
 
     def _nothing_(self):
         pass
@@ -76,6 +82,10 @@ class Valid:
         else:
             # Avoid user mistake with two time function set
             raise pexception.PytestembError("funcDestroy function already set")
+    
+    def set_fatal_mode(self, stop_case_run):
+        self.abort_fatal_mode = stop_case_run
+        
         
 
     def add_test_case(self, funcCase):
@@ -83,57 +93,77 @@ class Valid:
 
 
 
-
     def run_script(self):
-
         self.result.script_start({"name":self.name})
-        try:
-            # Create
+        # Create 
+        if self.create == self._nothing_:
+            pass
+        else:
             self.result.create_start({})
             self.run_try(self.create)
             self.result.create_stop({})
-            # Setup
+        # Setup
+        if self.setup == self._nothing_:
+            pass
+        else:
             self.result.setup_start({})
             self.run_try(self.setup)
             self.result.setup_stop({})
-            # Case
-            for acase in self.case :
-                name = acase.func_name
-                self.result.case_start({"name":name})
-                self.run_case(acase)
-                self.result.case_stop({"name":name})
-            # Cleanup
+        # Case
+        for acase in self.case :
+            name = acase.func_name
+            self.result.case_start({"name":name})
+            self.run_case(acase)
+            self.result.case_stop({"name":name})
+        # Cleanup
+        if self.cleanup == self._nothing_:
+            pass
+        else:
             self.result.cleanup_start({})
             self.run_try(self.cleanup)
             self.result.cleanup_stop({})
-            
-            # Destroy
+        # Destroy    
+        if self.destroy == self._nothing_:
+            pass
+        else:
             self.result.destroy_start({})
-            self.run_try(self.destroy)
+            self.run_try(self.destroy, force=True)
             self.result.destroy_stop({})
-            
-        except:
-            raise
+    
         self.result.script_stop({"name":self.name})
 
 
-    def run_try(self, func):
+    def run_try(self, func, force=False):
+        
+        if self.aborted and not force:
+            self.result.aborted({})
+            return
+
         try:
             func()
         except result.TestErrorFatal:
             pass
+        except result.TestAbort:
+            self.aborted = True                  
         except (Exception), (error):
             self.inspect_traceback(error)
 
+
     def run_case(self, case):
+        
+        if self.aborted or self.abort_fatal:
+            self.result.aborted({})
+            return
         try:
             case()
-            return True
         except result.TestErrorFatal:
-            return True
+            if self.abort_fatal_mode:
+                self.abort_fatal = True          
+        except result.TestAbort:
+            self.aborted = True        
         except (Exception), (error):
             self.inspect_traceback(error)
-            return False
+ 
 
 
 
@@ -150,7 +180,7 @@ class Valid:
                 stack[-1]["line"]      = copy.copy(traceback[index][2])
                 stack[-1]["function"]  = copy.copy(traceback[index][3])
                 stack[-1]["code"]      = copy.copy(traceback[index][4][0].strip("\n"))
-        except Exception, ex:
+        except Exception:
             pass
         finally:
             del traceback
@@ -160,6 +190,8 @@ class Valid:
         des["exception_class"]  = exception.__class__.__name__
 
         self.result.py_exception(des)
+
+
